added graphics for the main action buttons in /dashboard
